Fixed Bug: 1.3

+++ GrapholizeApp/src/main/java/util/Import/CompressedDot.java
@@ -15,7 +15,7 @@ public class CompressedDot {
 
     public int Twist;
 
-    public int Force;
+    public float Force;
 
     public int DotType;
     /*
